# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Biblioteca Reports System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  coordinator: 'coordinator',
  employee: 'employee',
} as const;

export interface User {
  id: number;
  username: string;
  name: string;
  role: UserRole;
}

export interface AuthResponse {
  user: User;
}

export interface Folder {
  id: number;
  ownerId: number;
  ownerName: string;
  createdAt: string;
}

export interface FolderWithCount {
  id: number;
  ownerId: number;
  ownerName: string;
  createdAt: string;
  reportCount: number;
}

export type ReportType = typeof ReportType[keyof typeof ReportType];


export const ReportType = {
  report: 'report',
  record: 'record',
} as const;

export interface Report {
  id: number;
  folderId: number;
  authorId: number;
  authorName: string;
  title: string;
  content: string;
  type: ReportType;
  createdAt: string;
  updatedAt: string;
}

export type ReportInputType = typeof ReportInputType[keyof typeof ReportInputType];


export const ReportInputType = {
  report: 'report',
  record: 'record',
} as const;

export interface ReportInput {
  /** @minLength 1 */
  title: string;
  content: string;
  type: ReportInputType;
}

export type ReportUpdateType = typeof ReportUpdateType[keyof typeof ReportUpdateType];


export const ReportUpdateType = {
  report: 'report',
  record: 'record',
} as const;

export interface ReportUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
  type?: ReportUpdateType;
}

export interface ActivityItem {
  reportId: number;
  title: string;
  authorName: string;
  type: string;
  updatedAt: string;
}

export interface DashboardStats {
  totalReports: number;
  totalRecords: number;
  totalFolders: number;
  recentActivity: ActivityItem[];
}