# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Biblioteca Reports System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: folders
    description: User folders
  - name: reports
    description: Reports and records
  - name: users
    description: Users

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout current user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (coordinator only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  /folders:
    get:
      operationId: listFolders
      tags: [folders]
      summary: List folders (coordinator sees all, employee sees own)
      responses:
        "200":
          description: List of folders with report counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FolderWithCount"

  /folders/{folderId}:
    get:
      operationId: getFolder
      tags: [folders]
      summary: Get a specific folder
      parameters:
        - name: folderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Folder details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /folders/{folderId}/reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List reports in a folder
      parameters:
        - name: folderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createReport
      tags: [reports]
      summary: Create a report in a folder
      parameters:
        - name: folderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

  /reports/{reportId}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a specific report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateReport
      tags: [reports]
      summary: Update a report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportUpdate"
      responses:
        "200":
          description: Updated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete a report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats:
    get:
      operationId: getStats
      tags: [reports]
      summary: Get dashboard statistics (coordinator only)
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"

    User:
      type: object
      required: [id, username, name, role]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [coordinator, employee]

    Folder:
      type: object
      required: [id, ownerId, ownerName, createdAt]
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        ownerName:
          type: string
        createdAt:
          type: string

    FolderWithCount:
      type: object
      required: [id, ownerId, ownerName, createdAt, reportCount]
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        ownerName:
          type: string
        createdAt:
          type: string
        reportCount:
          type: integer

    Report:
      type: object
      required: [id, folderId, authorId, authorName, title, content, type, createdAt, updatedAt]
      properties:
        id:
          type: integer
        folderId:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [report, record]
        createdAt:
          type: string
        updatedAt:
          type: string

    ReportInput:
      type: object
      required: [title, content, type]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        type:
          type: string
          enum: [report, record]

    ReportUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        type:
          type: string
          enum: [report, record]

    DashboardStats:
      type: object
      required: [totalReports, totalRecords, totalFolders, recentActivity]
      properties:
        totalReports:
          type: integer
        totalRecords:
          type: integer
        totalFolders:
          type: integer
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"

    ActivityItem:
      type: object
      required: [reportId, title, authorName, type, updatedAt]
      properties:
        reportId:
          type: integer
        title:
          type: string
        authorName:
          type: string
        type:
          type: string
        updatedAt:
          type: string