# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Biblioteca Reports System API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  reportId: number;
  title: string;
  authorName: string;
  type: string;
  updatedAt: string;
}