# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Biblioteca Reports System API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItem } from './activityItem';

export interface DashboardStats {
  totalReports: number;
  totalRecords: number;
  totalFolders: number;
  recentActivity: ActivityItem[];
}