# lib/api-zod/src/generated/types/folder.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Biblioteca Reports System API
 * OpenAPI spec version: 0.1.0
 */

export interface Folder {
  id: number;
  ownerId: number;
  ownerName: string;
  createdAt: string;
}