# lib/api-zod/src/generated/types/report.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Biblioteca Reports System API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportType } from './reportType';

export interface Report {
  id: number;
  folderId: number;
  authorId: number;
  authorName: string;
  title: string;
  content: string;
  type: ReportType;
  createdAt: string;
  updatedAt: string;
}