# lib/api-zod/src/generated/types/reportInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Biblioteca Reports System API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportInputType } from './reportInputType';

export interface ReportInput {
  /** @minLength 1 */
  title: string;
  content: string;
  type: ReportInputType;
}