# lib/api-zod/src/generated/types/reportUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Biblioteca Reports System API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportUpdateType } from './reportUpdateType';

export interface ReportUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
  type?: ReportUpdateType;
}